Use optimize_acqf for min_lcb evaluation
